Move code out of bin.cpp
