+++ lib/dataset/include/scipp/dataset/bin.h
@@ -4,7 +4,7 @@
 /// @author Simon Heybrock
 #pragma once
 
-#include <scipp/dataset/dataset.h>
+#include "scipp/dataset/dataset.h"
 
 namespace scipp::dataset {
